Create Final_Project.py
This is the .py code, I have tried it on pythonanywhere.com but it does not seem to work. There is an error when using the keys and secrets to access Twitter. I am thinking it may because I did not use Tweepy. I installed Tweepy at the beginning of program, but when trying the tweepy way of accessing developer account it did not work. So I tried this way and it still did not work. I think it may be because I am unsure how to assign twitter an API in the way it wants for the program to run.